Refactored rusty-roads (#24)
* Refactored rusty-roads

* added some comments and a box-query
